Update libwatchfish, use harbour- prefix
This changes the libwatchfish submodule to now point at Piggz updated
fork.
Changing to this new fork resolves various compilation issues due to an
incomplete implementation of various dbus services.

Additionally this changes the asteroidsyncserviced, when compiled with
"CONFIG+=starfish" to have "harbour-" prefixed to its name.

The reasoning behind this change, is that Jolla has a requirement for all
applications that are submittted to the offical store have this prefix.
https://harbour.jolla.com/faq#1.1.0