Fixes issue where no watches are added as AsteroidOS watches
By default the QStringList's Contains method is case sensitive.
This changes it so that when we check for the NOTIF_UUID it uses
a case insensentive comparison.

Previously there was a redeclaration of NOTIF_UUID that was in lowercase,
this was removed in a previous PR that tidied up quite a few redeclarations.
This unfortunately caused the uuid conparison to start failing
and AsteroidOS watches were nolonger recognised.