Handle platform differences in toplevel CMake file
This patch moves all platform difference handling to the top level
CMakeLists.txt file.  The top level defines two new CMake variables,
PLATFORM_SOURCE_DIR which is the path to platform-specific code relative
to the asteroidsyncservice subdirectory and DAEMON_PLATFORM_SOURCE_DIR
which is the path to platform-specific code relative to the
asteroidsyncserviced subdirectory.  Even though they currently have the
same value, they are two separate variables in anticipation of adding a
Linux desktop platform that uses D-Bus to start services from with
asteroidsyncservice (as in the current starfish code) and also uses
D-Bus in the daemon (as in the current ubunutu touch code).