Redefine QT wrapper template if needed
Somewhere before version 5.15.2 of Qt, the pointer wrapper used in the
QT_DECLARE_PRIVATE was defined to use p.data() explicitly instead of the
safer and more general p.operator->().  This provides the safer
definition only if Q_DECLARE_PRIVATE_SUPPORTS_UNIQUE_PTR is not set to 1.

Without this, older versions of Qt would fail to compile if a
std::unique_ptr was used instead of a raw pointer.