Test if redefinition of QT wrapper template needed
Somewhere before version 5.15.2 of Qt, the pointer wrapper used in the
QT_DECLARE_PRIVATE was defined to use p.data() explicitly instead of the
safer and more general p.operator->().  This provides the safer
definition only if Q_DECLARE_PRIVATE_SUPPORTS_UNIQUE_PTR is not set to 1.

Rather than checking for specific versions of Qt, this instead checks
directly for the problem with a small test program. It defines the CMake
variable Q_DECLARE_PRIVATE_SUPPORTS_UNIQUE_PTR = 1 if the code can compile
and there is no problem or = 0 if the problem is detected.