Remove SetVibration and add vibration to SendNotify
As the complementary change to the libasteroid library, this removes the
SetVibration call from the interface and adds an explicit vibration
parameter to SendNotify.  This has the effect that now all SendNotify
requests must explicitly set the vibration setting which makes more
sense than storing and not using a pre-set vibration value and
streamlines the interface.