Update SendNotify call for Ubuntu Touch to new API
Because the updated API now requires all callers to explicitly set the
vibration level for each SendNotify call, and because the libasteroid
library no longer supports getVibration(), the Ubuntu Touch platform now
always sets the vibration to "normal".  If a global setting is ever
created on that platform, this code could be modified to use that
setting rather than the hardcoded value.