Remove conversion in Watch::setScreenshotFileInfo
The Watch::setScreenshotFileInfo is passed a QString which is then
converted into a QFileInfo type within the function.  At the end, the
m_screenshotFileInfo variable was originally set with the fileInfo
string even though it's a QFileInfo type, forcing another conversion.
This eliminates that unnecessary conversion by simply using fInfo which
is already the correct type.

Signed-off-by: Ed Beroset <beroset@ieee.org>