Use open-meteo instead of openweathermap
Originally, the openweathermap application was used, but version 2.5
of that API will no longer work after June 2024, so now we use the
open-meteo API instead.

This parallels a change that was made in the asteroid-weatherfetch
source code and is largely a cut-and-paste copy with a few small changes
to use the existing interfaces.

Note that eventually, this code could go away, and the watch would
parse the JSON data itself, but that would require a change to the
bluetooth implementation on the watch.

Signed-off-by: Ed Beroset <beroset@ieee.org>